* You can open your backpack! * New Container entity, replaces ContainerItem. * ContainerGump now draws items (represented by ItemGumpling entities). * New Position3D class which replaces all instances of DrawPosition and TilePosition. * Internalized all entity Movement object calls. Entity now exposes a bare minimum of methods and variables necessary to determine position and movement. * Changed the way that Entities report their position to the World so they can be drawn. * All textures are now formated rgba8888, (formerly a mix of this and bgra5551).

+++ UltimaXNA/UltimaXNA/TileEngine/MouseOverList.cs
@@ -170,8 +170,8 @@ namespace UltimaXNA.TileEngine
             }
             else
             {
-                UInt16[] iPixel = new UInt16[1];
-                Texture.GetData<UInt16>(0,
+                uint[] iPixel = new uint[1];
+                Texture.GetData<uint>(0,
                     new Rectangle((int)mousePosition.X - (int)Position.X, (int)mousePosition.Y - (int)Position.Y, 1, 1), iPixel, 0, 1);
                 if (iPixel[0] != 0)
                     return true;
